requirements.txt
gunicorn==19.9.0
Flask==1.1.2
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
tabula-py==2.1.0
Werkzeug==1.0.1
